add default draw and touch, fix drag()
Fixes #116 - defaults for draw and touch make the zones actually show up
an interactive instead of the unusable invisible non-interactive
default.

Fixes #117 - drag() now works again for top-level zones, and also works
better for child zones, as we always keep the offset exactly, instead of
losing precision to integer conversions.

+++ src/vialab/SMT/Touch.java
@@ -328,7 +328,7 @@ public class Touch extends TuioCursor {
 	 * @param z
 	 * @return the x position of this Touch in local coordinates of the given zone
 	 */
-	public int getLocalX(Zone z){
+	public float getLocalX(Zone z){
 		return z.getLocalX(this);
 	}
 	
@@ -336,7 +336,7 @@ public class Touch extends TuioCursor {
 	 * @param z
 	 * @return the y position of this Touch in local coordinates of the given zone
 	 */
-	public int getLocalY(Zone z){
+	public float getLocalY(Zone z){
 		return z.getLocalY(this);
 	}	
 }
